% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_needed_breedings.R
\name{calculate_needed_breedings}
\alias{calculate_needed_breedings}
\title{Calculate the number of breedings needed to guarantee a fixed number of
offsprings with a certain probability}
\usage{
calculate_needed_breedings(
  confidence_p,
  effective_fertility_p,
  n_needed,
  litter_mean = NULL,
  offsprings_n_sample = NULL,
  litter_sd = 2.5,
  method = "textbook",
  calculation_type = ""
)
}
\arguments{
\item{confidence_p}{a number between 0 and 1}

\item{effective_fertility_p}{a number between 0 and 1}

\item{n_needed}{number of offsprings needed to be born}

\item{litter_mean}{the average litter size of the bred animals}

\item{litter_sd}{standart deviation of the litter size, equals 2.5
by default - as specified in the textbook}

\item{method}{the model used for the calculation, can be "textbook"}
}
\value{
The function returns the number of breedings
needed to guarantee \code{n_needed} offprings with probability no less than
\code{confidence_p}.
}
\description{
At the moment only one model is implemented (the textbook).
}
\examples{
calculate_needed_breedings(
    confidence_p=0.95,
    effective_fertility_p=0.6,
    n_needed = 30,
    litter_mean = 7,
    litter_sd = 2.5
    n_litters=10)

}
