% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectBorn.R
\name{expectBorn}
\alias{expectBorn}
\title{Expected total number of animals born in a breeding setup}
\usage{
expectBorn(breedingObj)
}
\arguments{
\item{breedingObj}{object created by multiGenotype function}
}
\value{

}
\description{
Expected total number of animals born in a breeding setup
}
