% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiGenotype.R
\name{multiGenotype}
\alias{multiGenotype}
\title{Optimal number of breedings, multiple genotype setup}
\usage{
multiGenotype(
  confidence_p = 0.9,
  birth_days = 3,
  genotypes_p = c(0.25, 0.5, 0.25),
  genotypes_N = c(0, 0, 10),
  sex_distribution = c("unimportant", "all one sex", "balanced"),
  litter_mean = NULL
)
}
\arguments{
\item{confidence_p}{}

\item{birth_days}{takes values 1,2,3 or 4: 
the desired offsprings should be born within `birth_days`; 
alternatibely, a number between 0 and 1 specifying the effective fertility 
of the animal}

\item{genotypes_p}{probabilities of a single offspring to have each of the respective genotypes, should sum to 1}

\item{genotypes_N}{required number of offsprings of the respective genotypes}

\item{sex_distribution}{takes values "unimportant", "all one sex", "balanced"; 
reflects the distribution of sexes among the offsprings: If "balanced", 
equal numbers of male and female offspring should be born for each genotype; 
if "all one sex", all of the required offsprings should be either male or female
if "unimportant", the offsprings of any sex suffice. 
Female and male offpsings are born with the same frequency (0.5).}

\item{litter_mean}{average number of offpring for one animal}
}
\value{

}
\description{
Calculates the smallest number of breedings needed to yield a desired number 
of offsprings of multiple genotypes with a given probability of success
}
