% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleGenotype.R
\name{singleGenotype}
\alias{singleGenotype}
\title{Calculates the smallest number of breeding needed to yield a desired number 
of offsprings of a single genotype with a given probability of success}
\usage{
singleGenotype(
  confidence_p = 0.9,
  birth_days,
  n_offsprings,
  litter_mean = NULL,
  method = "poisson"
)
}
\arguments{
\item{confidence_p}{}

\item{birth_days}{takes values 1,2,3 or 4: 
the desired offsprings should be born within `birth_days`; 
alternatibely, a number between 0 and 1 specifying the effective fertility 
of the animal}

\item{n_offsprings}{the desired number of offsprings}

\item{litter_mean}{average number of offpring for one animal}

\item{method}{takes values "festing" or "poisson"}
}
\value{

}
\description{
Calculates the smallest number of breeding needed to yield a desired number 
of offsprings of a single genotype with a given probability of success
}
